Modified the implementation to pass along Jwt Token strings and extract tenant context from it, instead of passing Jwt claims.

+++ metrics-java-lib/src/main/java/com/amazonaws/saas/metrics/builder/MetricBuilder.java
@@ -12,7 +12,7 @@ public class MetricBuilder {
         return this;
     }
 
-    public MetricBuilder withValue(Integer value) {
+    public MetricBuilder withValue(Long value) {
         this.metric.setValue(value);
         return this;
     }
